factor out subcomponents into own files
